requirements.txt
requests>=2.28,<3
pyyaml>=6.0,<7
urllib3>=1.26,<3
